debugged the previous render undefined issue
